Batch writes when sending.
Motivation:

Excessive traffic between NIO and Network.framework on the send path is
a source of performance limitations: as Network.framework is a resilient
library the calls into it are not as cheap as they could be. We can
therefore see some performance improvement if we change our code to
attempt to reduce the number of calls into Network.framework, especially
if we can do so without excessive byte copying.

Modifications:

- Updated the code to prefer to use DispatchData based writes when
available.

Result:

Faster write performance on newer OSes.